4 bit too
